ai-service/requirements.txt
fastapi
uvicorn
requests
python-dotenv
flask-cors
huggingface-hub
